Updated README with some project information and screenshots.
Updated lcd so that when it's switched on causes 4 cycles to elapse.
This is because blargg's lcd sync ROM on test 03 causes
a delay of 110 cycles (440 machine cycles). The store/load
instructions before/after the delay are both 12 machine cycles.
The last 8 cycles are used of the store as the lcd is then
switched on after the load from the first 4 cycles, and the first 4
cycles of the load instruction are added to the total before being
loaded into the A register. This only accounts for 452 machine cycles,
(440 + 8 + 4)
A completele scanline takes 456 cycles so I can only assume that the
extra 4 cycles come from switching on the lcd screen when the store
instruction is made to the LCDC register. This change also causes the
lcd sync test to pass.